+++ src/utils/menuFilter.ts
@@ -197,6 +197,13 @@ export const defaultMenuConfig: MenuItem[] = [
         icon: 'clipboard-document-list',
         requiredPermissions: ['edit_exams'],
       },
+      {
+        key: 'exams-edit',
+        label: 'Edit Exam',
+        path: '/exam/edit',
+        icon: 'clipboard-document-list',
+        requiredPermissions: ['edit_exams'],
+      },
     ],
   },
   {
